requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

ultralytics>=8.1.0

opencv-python>=4.8.0
pillow>=10.2.0
numpy>=1.26.0,<3
simple-lama-inpainting<=0.1.0
